refactor: 탈퇴 버튼 섹션 모듈화
